+++ flask_restplus/__about__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.10.1.dev'
+__version__ = '0.11.0.dev'
 __description__ = 'Fully featured framework for fast, easy and documented API development with Flask'