+++ flask_restplus/marshalling.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 
 from collections import OrderedDict
 from functools import wraps
+from six import iteritems
 
 from flask import request, current_app, has_app_context
 
@@ -10,7 +11,7 @@ from .mask import Mask, apply as apply_mask
 from .utils import unpack
 
 
-def marshal(data, fields, envelope=None, skip_none=False, mask=None):
+def marshal(data, fields, envelope=None, skip_none=False, mask=None, ordered=False):
     """Takes raw data (in the form of a dict, list, object) and a dict of
     fields to output and filters the data based on those fields.
 
@@ -22,6 +23,7 @@ def marshal(data, fields, envelope=None, skip_none=False, mask=None):
     :param bool skip_none: optional key will be used to eliminate fields
                            which value is None or the field's key not
                            exist in data
+    :param bool ordered: Wether or not to preserve order
 
 
     >>> from flask_restplus import fields, marshal
@@ -29,12 +31,21 @@ def marshal(data, fields, envelope=None, skip_none=False, mask=None):
     >>> mfields = { 'a': fields.Raw, 'c': fields.Raw, 'd': fields.Raw }
 
     >>> marshal(data, mfields)
-    OrderedDict([('a', 100), ('c', None), ('d', None)])
+    {'a': 100, 'c': None, 'd': None}
 
     >>> marshal(data, mfields, envelope='data')
-    OrderedDict([('data', OrderedDict([('a', 100), ('c', None), ('d', None)]))])
+    {'data': {'a': 100, 'c': None, 'd': None}}
 
     >>> marshal(data, mfields, skip_none=True)
+    {'a': 100}
+
+    >>> marshal(data, mfields, ordered=True)
+    OrderedDict([('a', 100), ('c', None), ('d', None)])
+
+    >>> marshal(data, mfields, envelope='data', ordered=True)
+    OrderedDict([('data', OrderedDict([('a', 100), ('c', None), ('d', None)]))])
+
+    >>> marshal(data, mfields, skip_none=True, ordered=True)
     OrderedDict([('a', 100)])
 
     """
@@ -50,22 +61,26 @@ def marshal(data, fields, envelope=None, skip_none=False, mask=None):
         fields = apply_mask(fields, mask, skip=True)
 
     if isinstance(data, (list, tuple)):
-        out = [marshal(d, fields, skip_none=skip_none) for d in data]
+        out = [marshal(d, fields, skip_none=skip_none, ordered=ordered) for d in data]
         if envelope:
-            out = OrderedDict([(envelope, out)])
+            out = OrderedDict([(envelope, out)]) if ordered else {envelope: out}
         return out
 
-    items = ((k, marshal(data, v, skip_none=skip_none) if isinstance(v, dict)
-              else make(v).output(k, data))
-             for k, v in fields.items())
+    items = (
+        (k, marshal(data, v, skip_none=skip_none, ordered=ordered)
+        if isinstance(v, dict)
+        else make(v).output(k, data, ordered=ordered))
+        for k, v in iteritems(fields)
+    )
 
     if skip_none:
-        items = ((k, v) for k, v in items if v is not None and v != OrderedDict())
+        items = ((k, v) for k, v in items
+                 if v is not None and v != OrderedDict() and v != {})
 
-    out = OrderedDict(items)
+    out = OrderedDict(items) if ordered else dict(items)
 
     if envelope:
-        out = OrderedDict([(envelope, out)])
+        out = OrderedDict([(envelope, out)]) if ordered else {envelope: out}
 
     return out
 
@@ -102,7 +117,7 @@ class marshal_with(object):
 
     see :meth:`flask_restplus.marshal`
     """
-    def __init__(self, fields, envelope=None, skip_none=False, mask=None):
+    def __init__(self, fields, envelope=None, skip_none=False, mask=None, ordered=False):
         """
         :param fields: a dict of whose keys will make up the final
                        serialized response output
@@ -112,6 +127,7 @@ class marshal_with(object):
         self.fields = fields
         self.envelope = envelope
         self.skip_none = skip_none
+        self.ordered = ordered
         self.mask = Mask(mask, skip=True)
 
     def __call__(self, f):
@@ -124,9 +140,13 @@ class marshal_with(object):
                 mask = request.headers.get(mask_header) or mask
             if isinstance(resp, tuple):
                 data, code, headers = unpack(resp)
-                return marshal(data, self.fields, self.envelope, self.skip_none, mask), code, headers
+                return (
+                    marshal(data, self.fields, self.envelope, self.skip_none, mask, self.ordered),
+                    code,
+                    headers
+                )
             else:
-                return marshal(resp, self.fields, self.envelope, self.skip_none, mask)
+                return marshal(resp, self.fields, self.envelope, self.skip_none, mask, self.ordered)
         return wrapper