+++ flask_restplus/namespace.py
@@ -5,11 +5,12 @@ import inspect
 import six
 import warnings
 
+from flask import request
 from flask.views import http_method_funcs
 
 from .errors import abort
 from .marshalling import marshal, marshal_with
-from .model import Model, SchemaModel
+from .model import Model, OrderedModel, SchemaModel
 from .reqparse import RequestParser
 from .utils import merge
 from ._http import HTTPStatus
@@ -26,9 +27,11 @@ class Namespace(object):
     :param str path: An optional prefix path. If not provided, prefix is ``/+name``
     :param list decorators: A list of decorators to apply to each resources
     :param bool validate: Whether or not to perform validation on this namespace
+    :param bool ordered: Whether or not to preserve order on models and marshalling
     :param Api api: an optional API to attache to the namespace
     '''
-    def __init__(self, name, description=None, path=None, decorators=None, validate=None, **kwargs):
+    def __init__(self, name, description=None, path=None, decorators=None, validate=None,
+            authorizations=None, ordered=False, **kwargs):
         self.name = name
         self.description = description
         self._path = path
@@ -41,6 +44,8 @@ class Namespace(object):
         self.resources = []
         self.error_handlers = {}
         self.default_error_handler = None
+        self.authorizations = authorizations
+        self.ordered = ordered
         self.apis = []
         if 'api' in kwargs:
             self.apis.append(kwargs['api'])
@@ -139,7 +144,8 @@ class Namespace(object):
 
         .. seealso:: :class:`Model`
         '''
-        model = Model(name, model, mask=mask)
+        cls = OrderedModel if self.ordered else Model
+        model = cls(name, model, mask=mask)
         model.__apidoc__.update(kwargs)
         return self.add_model(name, model)
 
@@ -229,7 +235,7 @@ class Namespace(object):
                 '__mask__': kwargs.get('mask', True),  # Mask values can't be determined outside app context
             }
             func.__apidoc__ = merge(getattr(func, '__apidoc__', {}), doc)
-            return marshal_with(fields, **kwargs)(func)
+            return marshal_with(fields, ordered=self.ordered, **kwargs)(func)
         return wrapper
 
     def marshal_list_with(self, fields, **kwargs):
@@ -309,6 +315,11 @@ class Namespace(object):
         for arg in args:
             kwargs.update(arg)
         return self.doc(vendor=kwargs)
+
+    @property
+    def payload(self):
+        '''Store the input payload in the current request context'''
+        return request.get_json()
 
 
 def unshortcut_params_description(data):