+++ flask_restplus/postman.py
@@ -4,12 +4,13 @@ from __future__ import unicode_literals, absolute_import
 from time import time
 from uuid import uuid5, NAMESPACE_URL
 
+from six import iteritems
 from six.moves.urllib.parse import urlencode
 
 
 def clean(data):
     '''Remove all keys where value is None'''
-    return dict((k, v) for k, v in data.items() if v is not None)
+    return dict((k, v) for k, v in iteritems(data) if v is not None)
 
 
 DEFAULT_VARS = {
@@ -54,15 +55,15 @@ class Request(object):
 
         # Add security headers if needed (global then local)
         for security in self.collection.api.__schema__.get('security', []):
-            for key, header in self.collection.apikeys.items():
+            for key, header in iteritems(self.collection.apikeys):
                 if key in security:
                     headers[header] = ''
         for security in self.operation.get('security', []):
-            for key, header in self.collection.apikeys.items():
+            for key, header in iteritems(self.collection.apikeys):
                 if key in security:
                     headers[header] = ''
 
-        lines = [':'.join(line) for line in headers.items()]
+        lines = [':'.join(line) for line in iteritems(headers)]
         return '\n'.join(lines)
 
     @property
@@ -97,7 +98,7 @@ class Request(object):
         params.update(dict((p['name'], p) for p in self.operation.get('parameters', [])))
         if not params:
             return url, None
-        for name, param in params.items():
+        for name, param in iteritems(params):
             if param['in'] == 'path':
                 url = url.replace('{%s}' % name, ':%s' % name)
                 path_vars[name] = DEFAULT_VARS.get(param['type'], '')
@@ -159,10 +160,10 @@ class PostmanCollectionV1(object):
                 'summary': 'The API Swagger specifications as JSON',
             })
         # Then iter over API paths and methods
-        for path, operations in self.api.__schema__['paths'].items():
+        for path, operations in iteritems(self.api.__schema__['paths']):
             path_params = operations.get('parameters', [])
 
-            for method, operation in operations.items():
+            for method, operation in iteritems(operations):
                 if method != 'parameters':
                     yield Request(self, path, path_params, method, operation)
 
@@ -175,7 +176,7 @@ class PostmanCollectionV1(object):
     def apikeys(self):
         return dict(
             (name, secdef['name'])
-            for name, secdef in self.api.__schema__.get('securityDefinitions').items()
+            for name, secdef in iteritems(self.api.__schema__.get('securityDefinitions'))
             if secdef.get('in') == 'header' and secdef.get('type') == 'apiKey'
         )