+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ Changelog
 Current
 -------
 
+- Nothing yet
+
+0.11.0 (2018-05-16)
+-------------------
+
+- Add authorizations parsing to namespace (:issue:`403`)
 - Add vendor extensions support (:issue:`97`)
 - :class:`~reqparse.RequestParser` arguments now support the ``split`` action
 - Ensure default boolean value as `False` works with :class:`~reqparse.RequestParser` (:issue:`199`)
@@ -19,6 +25,11 @@ Current
 - Add an optional key argument, ``skip_none``, in :func:`marshal_with` and :func:`marshal`
 - Fix masks not working correctly with Python 2.7 (:issue:`217`)
 - Fixed typos in doc/scaling
+- Add docs for `allow_null` and :class:`~fields.Nested`
+- Add Namespace.payload
+- **Breaking**: everything is unordered by default because ordering has a serious impact on performances:
+    - :class:`Api` and :class:`Namespace` now accept an optionnal ``ordered`` parameter
+    - :func:`marshal_with` and :func:`marshal` now accept an optionnal ``ordered`` parameter
 
 Breaking changes
 ~~~~~~~~~~~~~~~~